[FS] add all handlers layers

+++ xor-go/services/finances/internal/domain/filters.go
@@ -6,7 +6,7 @@ func CreateBankAccountFilter(
 	uuid *uuid.UUID,
 	accountId *uuid.UUID,
 	login *string,
-	funds *float64,
+	funds *float32,
 	status *string,
 ) BankAccountFilter {
 	return BankAccountFilter{
